+++ src/components/landing-page/HeroSectionNew.tsx
@@ -15,12 +15,14 @@ interface AssignmentHelpProps {
   changeInH: string;
   changeInP: string;
   isboy: boolean;
+  isReviewPage: boolean;
 }
 
 export default function HeroSectionNew({
   changeInH,
   changeInP,
   isboy,
+  isReviewPage,
 }: AssignmentHelpProps) {
   const [inputTypeCalen, setInputTypeCalen] = useState("text");
 
@@ -46,7 +48,16 @@ export default function HeroSectionNew({
       <div className="relative w-full h-screen flex justify-center items-center">
         <div className="right absolute bottom-16 left-[50%] ">
           <div>
-            {isboy ? (
+            {isReviewPage ? (
+              <Image
+                src="/static/images/review.png"
+                alt="review"
+                width={600}
+                height={600}
+                quality={100}
+                priority
+              />
+            ) : isboy ? (
               <Image
                 src="/static/images/girl.png"
                 alt="girl"
@@ -58,7 +69,7 @@ export default function HeroSectionNew({
             ) : (
               <Image
                 src="/static/images/boy.png"
-                alt="girl"
+                alt="boy"
                 width={1500}
                 height={1500}
                 quality={100}
@@ -66,6 +77,7 @@ export default function HeroSectionNew({
               />
             )}
           </div>
+
           <div className="absolute top-[15%] left-[73%] font-poppins bg-white rounded-xl shadow-xl px-4 py-8">
             {statsData.map((stat, index) => (
               <div
