+++ src/components/landing-page/faqs1.tsx
